CLTK-4 : Controller, Service 클래스 article, comment 분리
